FIX: many fixes

+++ src/main/java/com/example/toyshop/dto/product/ProductDetailDTO.java
@@ -11,13 +11,13 @@ public class ProductDetailDTO {
     private Long id;
     private String title;
     private Float price;
-    private Short amount;
+    private Integer amount;
     private String description;
     private Float rating;
     private CategoryProductDTO category;
     private List<ProductImageDTO> images;
     private List<ProductCommentDTO> comments;
     private Float new_price;
-    private Float discount;
-    private Long delivery_period;
+    private Short discount;
+    private Short delivery_period;
 }
